Refactor pymaid.fetch into subpackage
